refactor(server): split data_model associations.rs into per-association-type files
Addresses review feedback (AGENTS.md "split modules over ~400 lines"): the
combined associations.rs landed at 461 LOC. Break it into three cohesive
sibling files - classifications.rs (119), materials.rs (233), documents.rs
(109) - one per IfcRelAssociates* concern. The two shared helpers
read_logical / related_object_ids move into mod.rs as pub(super) (used across
all three). Pure code motion; mod.rs now 390. cargo test -p ifc-lite-server
green; clippy -D warnings clean.
